fix(): Allow http requests to be made from browser context
Raised in https://github.com/firecamp-dev/firecamp/issues/243

Previously, it was not possible to make a http request from the web client as
the request would be transparently upgraded to https. This was consequence of
an applied content security policy header with a directive to upgrade insecure
requests.

This change removes this CSP tag, meaning http requests are no longer forcibly
upgraded.